RCAL-1392 Refactor public vs private interface (#2314)
