+++ changes/2314.orientation.rst
@@ -1,0 +1,2 @@
+Refactor the orientation and engineering database code to make the public vs.
+private interface clear.
