+++ pyproject.toml
@@ -83,11 +83,11 @@ romancal = "romancal.stpipe.integration:get_steps"
 asn_from_list = "romancal.associations.asn_from_list:_cli"
 mk_skycell_asn_from_skycell_list= "romancal.associations.mk_skycell_asn_from_skycell_list:_cli"
 mk_skycell_list = "romancal.associations.mk_skycell_list:_cli"
-roman_pointing_summary = "romancal.scripts.pointing_summary:main"
-roman_set_telescope_pointing = "romancal.scripts.set_telescope_pointing:main"
-roman_set_velocity_aberration = "romancal.scripts.set_velocity_aberration:main"
+roman_pointing_summary = "romancal.scripts.pointing_summary:_main"
+roman_set_telescope_pointing = "romancal.scripts.set_telescope_pointing:_main"
+roman_set_velocity_aberration = "romancal.scripts.set_velocity_aberration:_main"
 roman_static_preview = "romancal.scripts.static_preview:command"
-roman_v1_calculate = "romancal.scripts.v1_calculate:main"
+roman_v1_calculate = "romancal.scripts.v1_calculate:_main"
 skycell_asn = "romancal.associations.skycell_asn:_cli"
 multiband_asn = "romancal.associations.multiband_asn:_cli"
