+++ romancal/lib/engdb/engdb_edp.py
@@ -4,13 +4,10 @@ import logging
 from os import getenv
 
 import numpy as np
-import requests
 from astropy.table import Table
 from astropy.time import Time
-from requests.adapters import HTTPAdapter, Retry
 
 from .engdb_lib import (
-    FORCE_STATUSES,
     RETRIES,
     TIMEOUT,
     EngdbABC,
@@ -18,10 +15,6 @@ from .engdb_lib import (
 )
 
 __all__ = ["EngdbEDP"]
-
-# Default MAST info.
-MAST_BASE_URL = "https://masttest.stsci.edu"
-API_URI = "edp/api/v0.1/mnemonics/spa/roman/data"
 
 # Configure logging
 logger = logging.getLogger(__name__)
@@ -84,7 +77,7 @@ class EngdbEDP(EngdbABC):
         logger.debug("kwargs not used by this service: %s", service_kwargs)
 
         try:
-            self.configure(environment, path_to_cc)
+            self._configure(environment, path_to_cc)
         except (FileNotFoundError, ModuleNotFoundError) as exception:
             raise RuntimeError(
                 f"Cannot instantiate EDP service with environment: {environment}, path_to_cc: {path_to_cc}"
@@ -92,32 +85,6 @@ class EngdbEDP(EngdbABC):
 
         # Check for basic aliveness.
         self.mr.get_mnemonic_metadata()
-
-    def configure(self, ops_env, path_to_cc):
-        """
-        Configure from parameters and environment.
-
-        Parameters
-        ----------
-        environment : ['dev', 'test', 'int', 'ops']
-            The operating environment currently working under.
-
-        path_to_cc : Path-like
-            Path to the kerberos keytab required to access the database.
-        """
-        self.ops_env = ops_env
-        self.path_to_cc = path_to_cc
-
-        # Create the mnemonic reader
-        from edp.mnemonics_reader import MnemonicsReader
-
-        self.mr = MnemonicsReader(
-            mission="roman", env=ops_env, access_type="FqA", path_to_cc=path_to_cc
-        )
-
-        # Get various timeout parameters
-        self.retries = int(getenv("ENG_RETRIES", RETRIES))
-        self.timeout = int(getenv("ENG_TIMEOUT", TIMEOUT))
 
     def get_meta(self, *kwargs):
         """
@@ -208,24 +175,31 @@ class EngdbEDP(EngdbABC):
 
         return results.collection
 
-    def set_session(self):
-        """Set up HTTP session."""
-        self._req = requests.Request(
-            method="GET",
-            url=self.base_url + API_URI,
-            headers={"Authorization": f"token {self.token}"},
+    def _configure(self, ops_env, path_to_cc):
+        """
+        Configure from parameters and environment.
+
+        Parameters
+        ----------
+        environment : ['dev', 'test', 'int', 'ops']
+            The operating environment currently working under.
+
+        path_to_cc : Path-like
+            Path to the kerberos keytab required to access the database.
+        """
+        self.ops_env = ops_env
+        self.path_to_cc = path_to_cc
+
+        # Create the mnemonic reader
+        from edp.mnemonics_reader import MnemonicsReader
+
+        self.mr = MnemonicsReader(
+            mission="roman", env=ops_env, access_type="FqA", path_to_cc=path_to_cc
         )
 
-        s = requests.Session()
-        retries = Retry(
-            total=self.retries,
-            backoff_factor=1.0,
-            status_forcelist=FORCE_STATUSES,
-            raise_on_status=True,
-        )
-        s.mount("https://", HTTPAdapter(max_retries=retries))
-
-        self._session = s
+        # Get various timeout parameters
+        self.retries = int(getenv("ENG_RETRIES", RETRIES))
+        self.timeout = int(getenv("ENG_TIMEOUT", TIMEOUT))
 
     def _get_records(
         self, mnemonic, starttime, endtime, time_format=None, **other_kwargs
