+++ romancal/orientation/_lib.py
@@ -1,0 +1,283 @@
+"""General library to support the orientation utilities"""
+
+import logging
+from collections import namedtuple
+from math import cos, sin
+
+import numpy as np
+
+__all__ = []
+
+# Setup logging
+logger = logging.getLogger(__name__)
+logger.addHandler(logging.NullHandler())
+DEBUG_FULL = logging.DEBUG - 1
+LOGLEVELS = [logging.INFO, logging.DEBUG, DEBUG_FULL]
+
+# Default and pre-defined matricies.
+# Conversion of the FGS Unified Frame (FGS) reference point from the V-Frame.
+# This is the pre-launch value, later to be refined and provided
+# in the SIAF.
+# Last update: 20260505 by T.Sohn
+FGS_DEFAULT_QUATERNION = np.array(
+    [-0.18596734175399293, 0.6837984564491885, -0.1800546332580956, 0.6822141509826322]
+)
+
+# Maximum absolute speed of the observatory. Used for sanity check is defined
+# as the sum of the absolute components of the velocity.
+MAX_OBSERVATORY_SPEED = 150
+
+# Common conversions.
+# Yes, these are available in a number of packages. However, they are explicitly defined
+# to provide consistency and, if any were to be changed, an explicit history of that change.
+
+# Conversion from seconds to MJD
+SECONDS2MJD = 1 / 24 / 60 / 60
+
+# Degree, radian, angle transformations
+R2D = 180.0 / np.pi
+D2R = np.pi / 180.0
+A2R = D2R / 3600.0
+R2A = 3600.0 * R2D
+PI2 = np.pi * 2.0
+
+# Potential errors generated by mnemonic retrival
+EXPECTED_ERRORS = (OSError, RuntimeError, ValueError)
+
+# WCS reference container
+WCSRef = namedtuple("WCSRef", ["ra", "dec", "pa", "s_region"])
+WCSRef.__new__.__defaults__ = (None, None, None, "")
+
+
+def angle_to_vector(alpha, delta):
+    """
+    Convert spherical angles to unit vector.
+
+    Parameters
+    ----------
+    alpha, delta : float
+        Spherical angles in radians.
+
+    Returns
+    -------
+    v : [float, float, float]
+        Unit vector.
+    """
+    v0 = cos(delta) * cos(alpha)
+    v1 = cos(delta) * sin(alpha)
+    v2 = sin(delta)
+
+    return [v0, v1, v2]
+
+
+def attitude_from_v1(vinfo):
+    """Calculate observatory attitude matrix based on V1 pointing
+
+    Return the attitude matrix used by `pysiaf.Aperture.set_attitude_matrix`
+    to enable transformations to/from the 'sky' frame.
+
+    Parameters
+    ----------
+    vinfo : `WCSRef`
+        The WCS information for V1, the boresight.
+
+    Returns
+    -------
+    attitude : np.array((3, 3), dtype=float)
+        The attitude matrix.
+    """
+    from pysiaf import Siaf
+    from pysiaf.utils.rotations import attitude_matrix
+
+    siaf = Siaf("roman")
+    boresight = siaf["BORESIGHT"]
+    v_refpoint = boresight.reference_point(to_frame="tel")
+    attitude = attitude_matrix(*v_refpoint, vinfo.ra, vinfo.dec, vinfo.pa)
+
+    return attitude
+
+
+def calc_position_angle(point, ref):
+    """
+    Calculate position angle from reference to point.
+
+        tan(pa) = cos(dec_r) * sin(ra_r - ra_p) / (sin(dec_r)cos(dec_p) - cos(dec_r)sin(dec_p)cos(ra_r-ra_p))
+
+    where::
+
+        pa : position angle
+        *_r : reference
+        *_p : point
+
+    Typically the reference is the V3 RA/DEC and point is the object RA/DEC.
+
+    Parameters
+    ----------
+    point : WCSRef
+        The POINT wcs parameters, in radians.
+
+    ref : WCSRef
+        The TARGET wcs parameters, in radians.
+
+    Returns
+    -------
+    point_pa : float
+      The POINT position angle, in radians
+    """
+    y = cos(ref.dec) * sin(ref.ra - point.ra)
+    x = sin(ref.dec) * cos(point.dec) - cos(ref.dec) * sin(point.dec) * cos(
+        ref.ra - point.ra
+    )
+    point_pa = np.arctan2(y, x)
+    if point_pa < 0:
+        point_pa += PI2
+    if point_pa >= PI2:
+        point_pa -= PI2
+
+    logger.debug("Given reference: %s, point: %s, then PA: %s", ref, point, point_pa)
+    return point_pa
+
+
+def hv_to_fgs(aperture_name, h, v):
+    """Convert HV frame to FGS frame
+
+    Source document: Innerspace Confluence: "Quaternion Transforms for Coarse Pointing WCS"
+    from the section "Derive M_fgstogsapp". The coordinates systems are based on
+    Eq. 7 of "Roman-STScI-000416 (Roman SOC FGS Algorithms)"
+
+    Parameters
+    ----------
+    aperture_name : str
+        The aperture in which the guide star is located.
+
+    h, v : float, float
+        The commanded coordinates in the HV frame.
+        Units are in pixels.
+
+    Returns
+    -------
+    fgs_x, fgs_y : float, float
+        The coordinates in the FGS reference frame in arcsec.
+    """
+    from pysiaf import Siaf
+
+    siaf = Siaf("roman")
+    aper = siaf[aperture_name]
+    aper_wfi_cen = siaf["WFI_CEN"]
+
+    # Location of GS. This is from Eqn. 7 of Roman-STScI-000416 (Roman SOC FGS Algorithms)
+    x, y = h + aper.XSciRef, aper.YSciRef - v
+    v2, v3 = aper.sci_to_tel(x, y)  # Convert sci to tel coords using pysiaf
+
+    # Convert tel to WFI_CEN idl coords using pysiaf.
+    # The only difference between FGS and WFI_CEN coordinates is that the Y-axis is flipped
+    x_wc, y_wc = aper_wfi_cen.tel_to_idl(v2, v3)
+    x_fgs, y_fgs = x_wc, -y_wc
+
+    return x_fgs, y_fgs
+
+
+def update_meta(model, t_pars, wcsinfo, vinfo, quality):
+    """Update model's meta info with the given pointing.
+
+    The following meta are update:
+    - meta.guide_star.hv_position
+    - meta.pointing.dec_v1
+    - meta.pointing.pa_aperture
+    - meta.pointing.pa_v3
+    - meta.pointing.quaternion
+    - meta.pointing.ra_v1
+    - meta.pointing.target_dec
+    - meta.pointing.target_ra
+    - meta.wcsinfo.dec_ref
+    - meta.wcsinfo.ra_ref
+    - meta.wcsinfo.roll_ref
+    - meta.wcsinfo.s_region
+
+    Parameters
+    ----------
+    model : `~roman.datamodels.DataModel`
+        The model to update. Updates are done in-place.
+
+    t_pars : `TransformParameters`
+        The transformation parameters. Parameters are updated during processing.
+
+    wcsinfo : `WCSRef``
+        The aperture-specific pointing.
+
+    vinfo : ``WCSRef`
+        The V1-specific pointing
+
+    quality : str
+        Indicator of the success of the pointing determination.
+    """
+    # Shortcuts to the meta blocks
+    gs = model.meta.guide_star
+    pm = model.meta.pointing
+    wm = model.meta.wcsinfo
+
+    # Setup SIAF info.
+    from pysiaf import Siaf
+
+    siaf = Siaf("roman")
+    aper = siaf[wm.aperture_name.upper()]
+
+    # Update pointing info
+
+    # Update SIAF-related meta
+    wm.v2_ref = aper.V2Ref
+    wm.v3_ref = aper.V3Ref
+    wm.v3yangle = aper.V3IdlYAngle
+    wm.vparity = aper.VIdlParity
+
+    # Update aperture-centric pointing, represented by `wcsinfo`
+    wm.dec_ref = wcsinfo.dec
+    wm.ra_ref = wcsinfo.ra
+    wm.roll_ref = wcsinfo.pa
+    wm.s_region = wcsinfo.s_region
+
+    # Update general pointing information, including V1.
+    pm.dec_v1 = vinfo.dec
+    pm.pa_aperture = wcsinfo.pa
+    pm.pa_v3 = vinfo.pa
+    pm.pointing_engineering_source = quality
+    pm.quaternion = tuple(t_pars.pointing.q)
+    pm.ra_v1 = vinfo.ra
+
+    # Update target's sky location.
+    attitude = attitude_from_v1(vinfo)
+    targ_app = siaf[model.meta.pointing.target_aperture]
+    targ_app.set_attitude_matrix(attitude)
+    skycoord = targ_app.reference_point(to_frame="sky")
+    pm.target_ra = skycoord[0]
+    pm.target_dec = skycoord[1]
+
+    # Update guidestar information
+    if t_pars.gscommanded is not None:
+        gs.hv_position = t_pars.gscommanded
+
+
+def vector_to_angle(v):
+    """
+    Return tuple of spherical angles from unit direction Vector.
+
+    The Direction Cosine Matrix (DCM) that provides the transformation of a
+    unit pointing vector defined in inertial frame (ECI J2000) coordinates to a
+    unit vector defined in the science aperture Ideal frame coordinates is
+    defined as.
+
+    Parameters
+    ----------
+    v : [v0, v1, v2]
+        Direction vector.
+
+    Returns
+    -------
+    alpha, delta : float, float
+        The spherical angles, in radians.
+    """
+    alpha = np.arctan2(v[1], v[0])
+    delta = np.arcsin(v[2])
+    if alpha < 0.0:
+        alpha += 2.0 * np.pi
+    return alpha, delta
