+++ romancal/orientation/tests/test_v1_calculate.py
@@ -7,8 +7,8 @@ import pytest
 from astropy.table import Table
 from roman_datamodels.datamodels import ScienceRawModel
 
+import romancal.orientation._v1_calculate as v1c
 import romancal.orientation.tests.test_set_telescope_pointing as tstp
-import romancal.orientation.v1_calculate as v1c
 
 # pysiaf is not a required dependency. If not present, ignore all this.
 pysiaf = pytest.importorskip("pysiaf")
